ArrayList /Java
ArrayList - (real world example) Imagine you have a magic backpack that can hold an unlimited number of toys. But this backpack is special: you can add toys, take out toys, and even check how many toys are inside, and it does all this super fast!
In java, an 'ArrayList' is like that magic backpack. it's a special kind of list that can even a bunch of items(like numbers,words,or even other objects).You can add items to 'ArrayList', remove them, or see what's inside at any time.
So, if you were coding a game, and you wanted to keep track of all the players' scores, you might use an 'ArrayList' to store all those scores. The best part is, the 'ArrayList' can grow or shrink depending on how many items you put in it, just like now your magic backpack can hold as many toys as you want.
-----------
In Java, an 'ArrayList' is a part of the Java Collection Framework and is resizable array implementation of the 'List' interface.